Fix CP2 implementation with regard to truncation of 32-bit MAC0/1/2/3 registers.
Up until now (including in the original Java and C versions) we have
been calculating MAC0/1/2/3 using an i64, so that we can detect the
extra bytes. Afterwards though, we have not been truncating it or
sign-extending from bit 31 upward.

This has meant that when calculating IR1/2/3 (for example) but more
generally when using the values again in the same opcode function, we
have extra bits in the upper 32 bits potentially, which throws off the
result. The new test caught these based on discrepancies between PhilPSX
and NOPSX for the RTPS opcode.

Already my new approach is paying dividends. The code (inside the flag
calculation function for MAC0/1/2/3) now properly handles truncation or
sign extension appropriately, and updates all uses of this function to
to store this newly returned value back into the previous holder.

With this, the test now passes properly, all values are exactly as they
are for NOPSX with the same register setup/input data. In addition, we
were also reading IR1 (which strictly speaking is a signed i16 inside a
32 bit register), which for the test data generates 0x8000.

When reading this, the test gets back 0xFFFF8000 as we store the sign
extended i64 version, truncating it to i32. This actually provides the
correct behaviour though, as when reading these register back to the
main CPU, NOPSX also sign extends. Test is now adjusted for this reason.