Adjust INTPL test and recompare with NOPSX.
Strictly speaking, for INTPL test we were putting the CODE value in
the least significant register byte rather than most significant. Test
still passed as we were also doing this in the fake BIOS file for NOPSX.
That said, CODE was just 0 for both NOPSX and our test because of this.

To make sure this value actually gets used in INTPL, we now put it in
the most significant byte, and re-run the test against the new NOPSX
values. It still passes.