Fix the bug uncovered by the NCDT test, and keep the extra NCDS test we added as well.
Essentially, the problem was that when calling handle_saturated_result
to calculate the value of IR1/2/3, we were passing a value >32 bits due
to us doing the calculation (for example):
((RFC << 12) - MAC1) >> 12

Normally, when we call handle_saturated_result, we just pass in results
of MAC1/2/3 directly, which is already truncated to 32 bits and sign
extended via its equivalent handle_unsaturated_result_and_truncate
function. In this case, we were directly passing the result above to it,
meaning the IR1/2/3 bounds checks were happening against this number
without correcting it first.

Now, I'm shadowing this value right at the start of the function, in
order to properly correct it in one place before performing the bounds
checks. With this fix, the NCDT test now passes, as does the new extra
test for NCDS that I added, and all existing tests as well.