Make minor tweaks to layout, define the motherboard, and introduce traits for this and the CPU. Finally, start defining the mechanisms to allow arbitrarily deep call chains where components can call each other (and back again). This is important to allow us to do what we can do in the C version.
In the C version, all the components just have raw pointers to each
other. That will not fly with my goal to use as little unsafe code as
possible in this version. The only way to make this work and be compile
time verified therefore is to use structures of component references and
transform them/pass them through as we go deeper down the call stack.

While this is quite verbose (and some macros would be useful in due
course) it works quite nicely, based on a test project I worked on for a
week or two when I had a chance - that is why this commit is so delayed!
That and me beginning to scratch the surface of how lifetimes work.